Ask the syntax where the printed text's inline comments stand
`string-quotes` was the last rule holding SCSS machinery of its own; it is syntax-blind now. The pair-diff that read the comments of a value off the two copies `postcss-scss` keeps — proving the pair in step by rewriting the comments it found and asking whether the raw comes back character for character — moves to `lib/preprocessor/findRewrittenCommentSpans/`, and the whole question folds into one contract method: `printedInlineComments(node, text, result)` answers off the pair while it is in step, off a scan of the text where the pair has drifted or the syntax keeps its comments in the text, and with nothing where the syntax spells none.

Two simplifications ride on what the seam already guarantees. The scss-first read of the value was spelled twice — once in `syntax.read` at the call site and once more inside the check — so the inner spelling goes. And the fix is written through `syntax.write` unconditionally: the position-mapped write into the raw copy was byte for byte what `write` regenerates, since the pair-diff itself proves `rewriteInlineComments(spelled) === raw` for a pair in step and a quote the rule fixes never stands inside a comment; for a pair out of step the write now repairs the raw from the fixed text rather than leaving it behind, which is the reading every other write of the plugin has had since the seam. `toRewrittenIndex` dies with the mapping.

Run over the change: `make verify`; `make oracles` against `origin/main`, all six oracles with no row added, removed or changed — the inline corpus of the control and twins oracles is largely a `string-quotes` corpus, which is the point of running it here.
