+++ scripts/check-break-readings.ts
@@ -77,6 +77,10 @@ const DEBT: Record<string, string[]> = {
 	],
 	"lib/rules/indentation/index.ts": [`target: \`\\n\`,`],
 	"lib/rules/max-empty-lines/index.ts": [`target: CRLF.test(rootString) ? \`\\r\\n\` : \`\\n\`,`],
+	"lib/preprocessor/findRewrittenCommentSpans/index.ts": [
+		`let lineBreakIndex = spelled.indexOf(\`\\n\`, spelledIndex)`,
+		`let rewrittenLineBreakIndex = rewritten.indexOf(\`\\n\`, rewrittenIndex)`,
+	],
 	"lib/rules/max-line-length/index.ts": [
 		`styleSearch({ source: rootString, target: [\`\\n\`], comments: \`check\` }, (match) => checkNewline(match))`,
 		`let nextNewlineIndex = rootString.indexOf(\`\\n\`, match.endIndex)`,
@@ -92,10 +96,6 @@ const DEBT: Record<string, string[]> = {
 	"lib/rules/selector-max-empty-lines/index.ts": [
 		`let violatedCRLFNewLinesRegex = new RegExp(\`(?:\\r\\n){\${maxAdjacentNewlines + 1},}\`, \`u\`)`,
 		`let violatedLFNewLinesRegex = new RegExp(\`\\n{\${maxAdjacentNewlines + 1},}\`, \`u\`)`,
-	],
-	"lib/rules/string-quotes/index.ts": [
-		`let lineBreakIndex = spelled.indexOf(\`\\n\`, spelledIndex)`,
-		`let rewrittenLineBreakIndex = rewritten.indexOf(\`\\n\`, rewrittenIndex)`,
 	],
 	"lib/rules/value-list-max-empty-lines/index.ts": [
 		`let violatedCRLFNewLinesRegex = new RegExp(\`(?:\\r\\n){\${maxAdjacentNewlines + 1},}\`, \`u\`)`,
